+++ typescript/src/kurtosis_core_rpc_api_bindings/api_container_service_pb.d.ts
@@ -737,3 +737,41 @@ export namespace ExecuteBulkCommandsArgs {
   }
 }
 
+export class GetServicesResponse extends jspb.Message {
+  getServiceIdsMap(): jspb.Map<string, boolean>;
+  clearServiceIdsMap(): void;
+  serializeBinary(): Uint8Array;
+  toObject(includeInstance?: boolean): GetServicesResponse.AsObject;
+  static toObject(includeInstance: boolean, msg: GetServicesResponse): GetServicesResponse.AsObject;
+  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
+  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
+  static serializeBinaryToWriter(message: GetServicesResponse, writer: jspb.BinaryWriter): void;
+  static deserializeBinary(bytes: Uint8Array): GetServicesResponse;
+  static deserializeBinaryFromReader(message: GetServicesResponse, reader: jspb.BinaryReader): GetServicesResponse;
+}
+
+export namespace GetServicesResponse {
+  export type AsObject = {
+    serviceIdsMap: Array<[string, boolean]>,
+  }
+}
+
+export class GetLambdasResponse extends jspb.Message {
+  getLambdaIdsMap(): jspb.Map<string, boolean>;
+  clearLambdaIdsMap(): void;
+  serializeBinary(): Uint8Array;
+  toObject(includeInstance?: boolean): GetLambdasResponse.AsObject;
+  static toObject(includeInstance: boolean, msg: GetLambdasResponse): GetLambdasResponse.AsObject;
+  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
+  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
+  static serializeBinaryToWriter(message: GetLambdasResponse, writer: jspb.BinaryWriter): void;
+  static deserializeBinary(bytes: Uint8Array): GetLambdasResponse;
+  static deserializeBinaryFromReader(message: GetLambdasResponse, reader: jspb.BinaryReader): GetLambdasResponse;
+}
+
+export namespace GetLambdasResponse {
+  export type AsObject = {
+    lambdaIdsMap: Array<[string, boolean]>,
+  }
+}
+