+++ typescript/src/lib/bulk_command_execution/v0_bulk_command_api/v0_bulk_commands.ts
@@ -2,6 +2,7 @@ import { LoadLambdaArgs, ExecCommandArgs, ExecuteBulkCommandsArgs, ExecuteLambda
 import { V0CommandTypeVisitor, V0CommandType } from "./v0_command_types";
 import { ok, err, Result } from "neverthrow";
 import * as protobuf from "google-protobuf";
+import * as google_protobuf_empty_pb from "google-protobuf/google/protobuf/empty_pb";
 
 // ====================================================================================================
 //                                   Command Arg Deserialization Visitor
@@ -147,6 +148,16 @@ class CmdArgDeserializingVisitor implements V0CommandTypeVisitor {
         const args: ExecuteBulkCommandsArgs = Object.assign(new ExecuteBulkCommandsArgs(), deserializationResult.value);
 
         this.deserializedCommandArgsPtr = args;
+        return ok(null);
+    }
+
+    public visitGetServices(): Result<null, Error> {
+        this.deserializedCommandArgsPtr = new google_protobuf_empty_pb.Empty()
+        return ok(null);
+    }
+
+    public visitGetLambdas(): Result<null, Error> {
+        this.deserializedCommandArgsPtr = new google_protobuf_empty_pb.Empty()
         return ok(null);
     }